USEN BAP-2 2차 작업본 적용 (L/R change, volume, BT/Aux Input select)

+++ AD85050.h
@@ -86,7 +86,7 @@ void AD85050_Amp_RAM_Set_Write(uint8_t uCount, uint8_t uData);
 
 void AD85050_Amp_Mode_Control(Audio_Output_Setting mode);
 
-void AD85050_Amp_Set_Cur_Volume_Level(uint8_t volume);
+void AD85050_Amp_Set_Cur_Volume_Level(uint32_t volume);
 
 Bool AD85050_Amp_Get_Cur_CLK_Status(void);
 void AD85050_Amp_Set_Default_Volume(void);
@@ -94,8 +94,8 @@ void AD85050_Amp_Set_Default_Volume(void);
 Bool AD85050_Amp_Detect_FS(Bool BInit); //BInit - TRUE : Amp Init / FALSE : Not need Amp Init //2022-10-17_2
 uint8_t AD85050_Amp_Detect_Fault(Bool Return_Val_Only); //2022-10-25 : FAULT PIN
 
-uint8_t AD85050_Amp_Get_Cur_Volume_Level(void);
-uint8_t AD85050_Amp_Volume_Set_with_Index(uint8_t Vol_Level, Bool Inverse, Bool Actual_Key); //Actual Key says this is not SSP or BLE communication. So, we need to send same key to Slave SPK
+uint32_t AD85050_Amp_Get_Cur_Volume_Level(void);
+uint32_t AD85050_Amp_Volume_Set_with_Index(uint32_t Vol_Level, Bool Inverse, Bool Actual_Key); //Actual Key says this is not SSP or BLE communication. So, we need to send same key to Slave SPK
 uint8_t AD85050_Amp_Get_Cur_Volume_Level_Inverse(void); //Start count from Min(0)
 
 void AD85050_Register_Read(void); //2022-10-25 : FAULT PIN
@@ -113,7 +113,7 @@ void AD85050_EQ_OnOff(Bool BEQ_On);
 void AD85050_DRC_OnOff(Bool BDRC_On);
 
 
-void AD85050_Amp_Volume_Register_Writing(uint8_t uVolumeLevel);
+void AD85050_Amp_Volume_Register_Writing(uint16_t uVolumeLevel);
 
 void AD85050_Dac_Volume_Set(Bool Aux_Mode); //2023-06-13_1 : Added Parameter
 #endif //AD85050_H
